User Registration

+++ base/views.py
@@ -16,6 +16,11 @@ from .forms import RoomForm
 #     {'id':3, 'name': 'Frontend developers'},
 
 def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -33,12 +38,16 @@ def loginPage(request):
         else:
             messages.error(request, 'Username OR password does not exist')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    page = 'register'
+    return render(request, 'base/login_register.html')
 
 def home(request):
     rooms = Room.objects.all()
@@ -69,7 +78,7 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
-    if request.user != room.user:
+    if request.user != room.host:
         return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
@@ -84,6 +93,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
